test(kueue): split the disruption suite under the file-size guard
CI was 18 lanes red from ONE root cause, and it was not the rustls
dev-dependency: `cargo-deny check` passed in the same job. `Server Guards`
failed at its next step, `Check changed Rust file sizes`:

  FAIL server/crates/djinn-k8s/tests/kueue_disruption_conformance.rs
       (1680 lines, 68924 bytes)
  Found 1 oversized file(s) under MAX_LINES=1500 / MAX_BYTES=51200.

Everything else in that run is the fail-fast cascade -- Server Clippy, the
eight test shards and sqlx Freshness produced no failure output of their own.

Two changes, in that order:

1. The module doc reproduced the whole PR body. Trimmed to a summary of the
   four live findings with the load-bearing "why" left inline at the code each
   one constrains; the narrative lives in the PR body and the previous commit
   message, and three copies of it was two too many.

2. That alone could not fit: comment blocks of >= 6 lines were only 16KB of
   the 66KB, so the file was large in CODE. Split into two test TARGETS over a
   shared helper module, which is what the guard is asking for:

     tests/kueue_disruption/mod.rs         876 lines / 32401 B  helpers
     tests/kueue_disruption_conformance.rs 526 lines / 23924 B  guards + AC1/AC2
     tests/kueue_disruption_governor.rs    267 lines / 11592 B  AC3/AC5

   `scripts/check-file-size.sh --files-from-stdin` now reports 0 oversized
   files across all three. The `djinn:allow-oversize` marker was deliberately
   NOT used: it is for source that genuinely cannot be divided, and this could.

`#![allow(dead_code)]` in the shared module is load-bearing rather than lazy —
it is compiled separately into each binary, so helpers the sibling target uses
are genuinely unused in this one and `-D warnings` would fail the lane for
code that is exercised.

No assertion changed. All four hermetic guards still pass and the three live
tests keep their #[ignore] + DJINN_TEST_KUEUE_CLUSTER=1 gating.
